requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --group dev -o requirements.txt
allure-pytest==2.15.2
    # via saucedemo-auth-automation (pyproject.toml:dev)
allure-python-commons==2.15.2
    # via allure-pytest
attrs==25.4.0
    # via allure-python-commons
certifi==2025.11.12
    # via requests
charset-normalizer==3.4.4
    # via requests
exceptiongroup==1.3.1
    # via pytest
greenlet==3.3.0
    # via playwright
idna==3.11
    # via requests
iniconfig==2.3.0
    # via pytest
loguru==0.7.3
    # via saucedemo-auth-automation (pyproject.toml)
packaging==25.0
    # via pytest
playwright==1.57.0
    # via
    #   saucedemo-auth-automation (pyproject.toml)
    #   pytest-playwright
pluggy==1.6.0
    # via
    #   allure-python-commons
    #   pytest
pyee==13.0.0
    # via playwright
pygments==2.19.2
    # via pytest
pytest==9.0.2
    # via
    #   saucedemo-auth-automation (pyproject.toml)
    #   allure-pytest
    #   pytest-base-url
    #   pytest-playwright
pytest-base-url==2.1.0
    # via pytest-playwright
pytest-playwright==0.7.2
    # via saucedemo-auth-automation (pyproject.toml)
python-slugify==8.0.4
    # via pytest-playwright
requests==2.32.5
    # via pytest-base-url
text-unidecode==1.3
    # via python-slugify
tomli==2.3.0
    # via pytest
typing-extensions==4.15.0
    # via
    #   exceptiongroup
    #   pyee
urllib3==2.6.2
    # via requests
